+++ src/db_value.rs
@@ -47,7 +47,7 @@ use crate::row::ColumnData;
 use crate::row::DbRow;
 use crate::rowid::Rowid;
 use crate::statement::CachedStatement;
-use crate::statement::StatementHolder;
+use crate::statement::Statement;
 use crate::utils;
 use crate::vector::Vector;
 use crate::write_buffer::ToBuf;
@@ -81,12 +81,9 @@ impl PendingDbValue {
     /// Converts pending data into a public value with a client reference.
     pub(crate) fn into_db_value(self, client_ref: &ClientRef) -> DbValue {
         match self {
-            PendingDbValue::Cursor(statement) => {
-                DbValue::Cursor(Box::new(Cursor::new(StatementHolder::new(
-                    client_ref.clone(),
-                    statement,
-                ))))
-            }
+            PendingDbValue::Cursor(statement) => DbValue::Cursor(Box::new(
+                Cursor::new(Statement::new(client_ref.clone(), statement)),
+            )),
             PendingDbValue::Lob(data) => {
                 DbValue::Lob(Lob::new(client_ref.clone(), data))
             }